chore: simplify payout tokens handling (#2665)
* chore: simplify payout tokens handling

* chore: simplify payout tokens handling

* chore: simplify payout tokens handling
