+++ packages/round-manager/src/features/api/application.ts
@@ -1,4 +1,4 @@
-import { fetchFromIPFS, PayoutToken, pinToIPFS } from "./utils";
+import { fetchFromIPFS, pinToIPFS } from "./utils";
 import {
   AppStatus,
   GrantApplication,
@@ -19,6 +19,7 @@ import { Signer } from "@ethersproject/abstract-signer";
 import { Web3Provider } from "@ethersproject/providers";
 import { graphql_fetch } from "common";
 import { DirectPayoutStrategy__factory } from "../../types/generated/typechain";
+import { PayoutToken } from "./payoutTokens";
 // import { verifyApplicationMetadata } from "common/src/verification";
 // import { fetchProjectOwners } from "common/src/registry";
